+++ neoforge/src/main/java/com/github/mystery2099/voxlib/neoforge/VoxLibNeoForgeClient.java
@@ -1,22 +1,27 @@
-package com.github.mystery2099.voxlib.forge;
+package com.github.mystery2099.voxlib.neoforge;
 
 import com.github.mystery2099.voxlib.config.VoxLibConfig;
 import com.github.mystery2099.voxlib.debug.VoxelShapeDebugClient;
 import net.minecraft.client.Minecraft;
-import net.minecraftforge.client.ConfigScreenHandler;
-import net.minecraftforge.client.event.RenderHighlightEvent;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.fml.ModLoadingContext;
-import net.minecraftforge.fml.loading.FMLPaths;
+import net.minecraft.client.gui.screens.Screen;
+import net.neoforged.fml.ModContainer;
+import net.neoforged.fml.loading.FMLPaths;
+import net.neoforged.neoforge.client.event.RenderHighlightEvent;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+import net.neoforged.neoforge.common.NeoForge;
 
-final class VoxLibForgeClient {
-    static void initialize() {
+final class VoxLibNeoForgeClient {
+    static void initialize(ModContainer modContainer) {
         VoxLibConfig.Companion.initialize(FMLPaths.CONFIGDIR.get());
         VoxelShapeDebugClient.INSTANCE.initialize();
-        ModLoadingContext.get().registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class,
-            () -> new ConfigScreenHandler.ConfigScreenFactory(
-                (minecraft, parent) -> VoxLibConfig.Companion.createConfigScreen(parent)));
-        MinecraftForge.EVENT_BUS.addListener(VoxLibForgeClient::renderBlockHighlight);
+        modContainer.registerExtensionPoint(IConfigScreenFactory.class,
+            new IConfigScreenFactory() {
+                @Override
+                public Screen createScreen(Minecraft minecraft, Screen parent) {
+                    return VoxLibConfig.Companion.createConfigScreen(parent);
+                }
+            });
+        NeoForge.EVENT_BUS.addListener(VoxLibNeoForgeClient::renderBlockHighlight);
     }
 
     private static void renderBlockHighlight(RenderHighlightEvent.Block event) {